pyproject.toml
[project]
name = "vqa-implementations"
version = "0.1.0"
description = "VQA implementation with 4 different approaches: Original VQA, SAN, BUTD, MCAN"

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pytorch-lightning>=2.0.0",
    "torchmetrics>=1.0.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "tqdm>=4.64.0",
    "h5py>=3.7.0",
    "nltk>=3.7",
    "spacy>=3.4.0",
    "transformers>=4.20.0",
    "opencv-python>=4.6.0",
    "albumentations>=1.3.0",
    "wandb>=0.13.0",
    "tensorboard>=2.10.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["models", "datasets", "utils", "scripts"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
allow_redefinition = true
check_untyped_defs = false
no_implicit_optional = false
show_error_codes = true
 